+++ app/src/main/java/com/academy/android/domain/use_cases/GetLikesCountForChatIdUseCase.kt
@@ -1,4 +1,4 @@
-package com.academy.android.domain.interactors
+package com.academy.android.domain.use_cases
 
 import com.academy.android.domain.repositories.NewsRepository
 import kotlinx.coroutines.flow.Flow
