+++ app/src/main/java/com/academy/android/domain/use_cases/GetPassedNewsUseCase.kt
@@ -1,4 +1,4 @@
-package com.academy.android.domain.interactors
+package com.academy.android.domain.use_cases
 
 import com.academy.android.domain.models.News
 import com.academy.android.domain.repositories.NewsRepository
