+++ app/src/main/java/com/academy/android/ui/news/NewsViewModel.kt
@@ -1,11 +1,11 @@
 package com.academy.android.ui.news
 
 import androidx.lifecycle.ViewModel
-import com.academy.android.domain.interactors.GetFeaturedNewsUseCase
-import com.academy.android.domain.interactors.GetLikesCountForChatIdUseCase
-import com.academy.android.domain.interactors.GetMessagesCountForChatIdUseCase
-import com.academy.android.domain.interactors.GetPassedNewsUseCase
-import com.academy.android.domain.interactors.NewsLikesInteractor
+import com.academy.android.domain.use_cases.GetFeaturedNewsUseCase
+import com.academy.android.domain.use_cases.GetLikesCountForChatIdUseCase
+import com.academy.android.domain.use_cases.GetMessagesCountForChatIdUseCase
+import com.academy.android.domain.use_cases.GetPassedNewsUseCase
+import com.academy.android.domain.use_cases.NewsLikesUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
@@ -20,7 +20,7 @@ class NewsViewModel @Inject constructor(
     getPassedNewsUseCase: GetPassedNewsUseCase,
     private val getMessagesCountForChatIdUseCase: GetMessagesCountForChatIdUseCase,
     private val getLikesCountForChatIdUseCase: GetLikesCountForChatIdUseCase,
-    private val newsLikesInteractor: NewsLikesInteractor,
+    private val newsLikesUseCase: NewsLikesUseCase,
 ) : ViewModel() {
 
     private val filterState = MutableStateFlow(true)
@@ -39,10 +39,10 @@ class NewsViewModel @Inject constructor(
         }
 
     fun getIsLiked(id: Long): Boolean =
-        newsLikesInteractor.getIsLikedForNewsId(id)
+        newsLikesUseCase.getIsLikedForNewsId(id)
 
     fun handleLike(id: Long, isLiked: Boolean): Boolean =
-        newsLikesInteractor.updateLikedForNewsId(id, isLiked)
+        newsLikesUseCase.updateLikedForNewsId(id, isLiked)
 
     fun getChatMessagesCount(id: Long): Flow<Int> =
         getMessagesCountForChatIdUseCase(id).flowOn(Dispatchers.IO)
